server/requirements.txt
pymongo==4.3.2
python-dotenv==0.21.0